// src/app/api/model/businessName.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Address } from './address';


export interface BusinessName {
    id?: number;
    name?: string;
    address?: Array<Address>;
}