// src/app/api/model/user.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AccountSeller } from './accountSeller';


export interface User {
    userId?: number;
    email?: string;
    password?: string;
    repeatPassword?: string;
    token?: string;
    role?: string;
    accountSeller?: AccountSeller;
}